Move logic from controller and change url
It appears that there is a change being made to the application architecture, specifically the logic that was previously in the OrdersController is being moved to the UserService and OrderService. Additionally, there is a fix being made to the URL links on the all_elements.html file in the /admin directory.